+++ src/components/Form/Form.js
@@ -4,7 +4,7 @@ const Form = () => (
   <form className="d-flex justify-content-between">
     <input className="input-title" placeholder="Book title" />
     <select className="category">
-      <option selected>Category</option>
+      <option defaultValue={' '}>Category</option>
       {' '}
     </select>
     <button className="submit" type="submit">SUBMIT</button>
